+++ MbSoftlab.TemplateEngine.Core.Demo/Address.cs
@@ -1,0 +1,8 @@
+﻿namespace MbSoftLab.TemplateEngine.Core.Demo
+{
+    public class Address
+    {
+        public string Street { get; set; }
+        public string PostCode { get; set; }
+    }
+}
