+++ MbSoftlab.TemplateEngine.Core.Demo/Order.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace MbSoftLab.TemplateEngine.Core.Demo
+{
+    public class Order
+    {
+        public int Id { get; set; }
+        public List<Product> Products { get; set; }
+    }
+}
