+++ MbSoftlab.TemplateEngine.Core.Demo/Product.cs
@@ -1,0 +1,8 @@
+﻿namespace MbSoftLab.TemplateEngine.Core.Demo
+{
+    public class Product
+    {
+        public string ProductName { get; set; }
+        public float Price { get; set; }
+    }
+}
